+++ docs-src/src/actions/config/index.jsx
@@ -107,6 +107,21 @@ export const setPythonConfig = config => (
       `LBASGroup3Nodes: ${combatLBASGroup3Nodes}`,
       `MiscOptions: ${combatOptions.join(',')}`,
       '',
+      '[ShipSwitcher]',
+      `Enabled: ${configTemp.shipSwitcherEnabled}`,
+      `Slot1Criteria: ${configTemp.shipSwitcherSlot1Criteria}`,
+      `Slot1Ships: ${configTemp.shipSwitcherSlot1Ships}`,
+      `Slot2Criteria: ${configTemp.shipSwitcherSlot2Criteria}`,
+      `Slot2Ships: ${configTemp.shipSwitcherSlot2Ships}`,
+      `Slot3Criteria: ${configTemp.shipSwitcherSlot3Criteria}`,
+      `Slot3Ships: ${configTemp.shipSwitcherSlot3Ships}`,
+      `Slot4Criteria: ${configTemp.shipSwitcherSlot4Criteria}`,
+      `Slot4Ships: ${configTemp.shipSwitcherSlot4Ships}`,
+      `Slot5Criteria: ${configTemp.shipSwitcherSlot5Criteria}`,
+      `Slot5Ships: ${configTemp.shipSwitcherSlot5Ships}`,
+      `Slot6Criteria: ${configTemp.shipSwitcherSlot6Criteria}`,
+      `Slot6Ships: ${configTemp.shipSwitcherSlot6Ships}`,
+      '',
       '[Quests]',
       `Enabled: ${configTemp.questsEnabled}`,
     ]
