+++ docs-src/src/localizations/en.jsx
@@ -68,6 +68,18 @@ export const en = {
     combatReserveDocks: 'Reserve Docks',
     combatPortCheck: 'Port Check',
     combatMedalStop: 'Medal Stop',
+    shipSwitcherHeader: 'Ship Switcher',
+    shipSwitcherSlot1: 'Slot 1',
+    shipSwitcherSlot2: 'Slot 2',
+    shipSwitcherSlot3: 'Slot 3',
+    shipSwitcherSlot4: 'Slot 4',
+    shipSwitcherSlot5: 'Slot 5',
+    shipSwitcherSlot6: 'Slot 6',
+    shipSwitcherCriteria: 'Switch Criteria',
+    shipSwitcherCriteriaDamage: 'Damage',
+    shipSwitcherCriteriaFatigue: 'Fatigue',
+    shipSwitcherCriteriaSparkle: 'Sparkle',
+    shipSwitcher: '',
     questsHeader: 'Quests',
     configHeader: 'Config',
     configLoad: 'Load',
