+++ docs-src/src/reducers/config/config.jsx
@@ -43,6 +43,19 @@ const jsonConfigDefaults = {
   combatOptionReserveDocks: false,
   combatOptionPortCheck: false,
   combatOptionMedalStop: false,
+  shipSwitcherEnabled: false,
+  shipSwitcherSlot1Criteria: null,
+  shipSwitcherSlot1Ships: null,
+  shipSwitcherSlot2Criteria: null,
+  shipSwitcherSlot2Ships: null,
+  shipSwitcherSlot3Criteria: null,
+  shipSwitcherSlot3Ships: null,
+  shipSwitcherSlot4Criteria: null,
+  shipSwitcherSlot4Ships: null,
+  shipSwitcherSlot5Criteria: null,
+  shipSwitcherSlot5Ships: null,
+  shipSwitcherSlot6Criteria: null,
+  shipSwitcherSlot6Ships: null,
   questsEnabled: true,
 }
