+++ package.json
@@ -31,14 +31,14 @@
     "css-loader": "^0.28.7",
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "file-loader": "^1.1.5",
-    "material-ui": "^1.0.0-beta.18",
+    "material-ui": "^1.0.0-beta.34",
     "material-ui-time-picker": "^0.1.2",
     "mdi": "^2.0.46",
     "mdi-material-ui": "^2.0.1",
     "path": "^0.12.7",
     "prop-types": "^15.6.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-dropzone": "^4.2.3",
     "react-redux": "^5.0.6",
     "react-select": "^1.0.0-rc.10",
